backend/requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
attrs==23.2.0
binaryornot==0.4.4
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==1.7.3
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fastapi-cache2==0.2.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
jinja2-time==0.2.0
manage-fastapi==1.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pendulum==3.0.0
poyo==0.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pydantic==1.10.14
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
questionary==1.10.0
redis==4.6.0
regex==2023.12.25
requests==2.31.0
ruff==0.3.2
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
sympy==1.12
text-unidecode==1.3
time-machine==2.14.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typer==0.4.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
wcwidth==0.2.13
yarl==1.9.4
